Got the Adhan library hooked up. Works with settings. Need to flush out swiftdata.
issues where changing anything in viewmodel re-renders the view prayertimesview so any navigationlinks we are in become interrupted since the originating view changed.
this is bc we injected prayerviewmodel as an environment model at global level.
we used to have the model injected as a stateobject managed by only prayertimesview. this was good because really only he needed acesss.
so im trying to decouple the dependency of viewmodel from prayertimesview and make it load its own set of prayers from context instead... then none of the views will be rendered from viewmodel. idek if this will work but gonna try it out.

experimenting and learning 🎨😌