Wow, that was a whirlwind
I learned a huge lesson today on observable objects (really good video i learned from: https://youtu.be/RvzJLekIjRs?si=jgMMtMnddhgshey6)

(Firstly, recall that in SwiftUI whenever a State changes it redraws the whole view. This means that everything in its body is thrown out, and redrawn.)

So I basically learned that view models are responsible for all the logic and the intermediary between data and a view.

Now, when you want to access of your model within view, you have to instantiate as an object. There are different flavors of objects:
1. @StateObject. (think a state var)
2. @ObservableObject. (think a binding var)
3. Environment object.

- StateObject is the source of truth. These do not get redrawn / recreated. They stay "persistent" throughout a view.
- ObservedObjects basically look to the model or the "binded" / passed in stateobject. If not bound to a source of truth, they get thrown away with the view and get redrawn. Aka the model always gets reinitialized -- i tested this with a print statement at one point lol(So you use this when passing it to other child views where StateObject was instantiated)

BIG NO NO: having stateobjects of same type in childviews. There should be ONE source of truth

- @EnvironmentObjects are a way for us to be able to bypass having to pass the objects as arguments into childviews. (Still needs a source of truth as they can not be initialized. They have to be from a @StateObject -- typically at root view)
   > childviews can access this AS LONG AS its been injected in the hierarchy at a parent view
   > https://youtu.be/iis7YSVIawg?si=zS2ApybkG7sBRryy

Now, I understand all that. Using that I did it properly, I tried many methods, but the two that I tried, which were correct were:
1. PrayerTimesView hold StateObject  → Children ObservedObject
   > Details: Passing context from ShukrApp.swift (main daddy) into prayerTimesView as argument. Using it to Initialize @StateObject of prayerViewModel. Passing that as @ObservedObject to child view
2.  StateObject @ Root → Injected with .environmentObject() on daddy view.
   > now everyone else just uses @EnvironmentObject var viewModel: PrayerViewModel.

Cool both of those should work in theory.... BUT THEY DIDNTTTT
When navigating to lets say the settings page, pressing a button that dealed with viewmodel, it just kicked me out of the page.
I kept getting error "Trying to pop to a missing destination - SwiftUICore/Logging.swift:84 - please file a bug report."
Even when clicking on the back button, it works the first time, and didnt the next time.
Then i fixed it in simulator. i kept moving forward coding... but then the real device was having the same issue...
I figured the issue is basically he doesn't know where to go back in his navigation stack. So that's when I figured that, even though we were using these object variables, there was something getting thrown out and re-drawn (especially since when we interacted with SettingsView buttons that interacted with viewmodel)

Researched it, and I found these:
- https://arc.net/l/quote/nqmszuxn
- https://forums.developer.apple.com/forums/thread/653367
TLDR; NavigationViews placed inside of TabViews don't like environment variable. So yes it probably was the object thing but also this at the same time. Cuz i would get the error sometimes and not at others.

my work around:

```
struct shukrApp: App {

    @StateObject var sharedState = SharedStateClass()
    @StateObject var prayerViewModel: PrayerViewModel // Add ViewModel here
    @State private var globalLocationManager = GlobalLocationManager()

    var sharedModelContainer: ModelContainer = {
        ...
    }()

    init() {
        let context = sharedModelContainer.mainContext
        _prayerViewModel = StateObject(wrappedValue: PrayerViewModel(context: context))
    }

    var body: some Scene {

            WindowGroup {

                    if globalLocationManager.isAuthorized{
                        TabView (selection: $sharedState.selectedViewPage){ /// <-- tabview on the outside
                            HistoryPageView()
                                .tag(0)

                            NavigationView{ /// <-- put navigationView AS A TAB!!!
                                PrayerTimesView()
                            }
                            .environmentObject(prayerViewModel) /// <-- injecting on the navigationview
                            .tag(1)

                            DuaPageView()
                                .tag(2)

                        }
                        .tabViewStyle(DefaultTabViewStyle()) // for testing
                    }
            }
            .modelContainer(sharedModelContainer)
            .environment(globalLocationManager)
            .environmentObject(sharedState) // Inject shared state into the environment (Global access point for `sharedState`)
    }
```

Now homeboy works on Simulator and on real device

The only reason I even use TabView was because I wanted sliding navigation, but I think I'm gonna let that go. Then we can just make the whole thing a NavigationView and have the buttons in the topBar be of NavigationLink